BDOG-1215 Refactor commonality in AuditFilters
