+++ bootstrap-backend-play-26/src/main/scala/uk/gov/hmrc/play/bootstrap/backend/filters/BackendAuditFilter.scala
@@ -16,151 +16,30 @@
 
 package uk.gov.hmrc.play.bootstrap.backend.filters
 
-import akka.stream._
-import akka.stream.scaladsl.{Flow, Keep, Sink, Source}
-import akka.util.ByteString
+import akka.stream.Materializer
 import javax.inject.Inject
-import play.api.Logger
-import play.api.http.HttpEntity
-import play.api.http.HttpEntity.Streamed
-import play.api.libs.streams.Accumulator
-import play.api.mvc.{Result, _}
-import play.api.routing.Router.Attrs
+import play.api.mvc.{RequestHeader, ResponseHeader, Result}
 import uk.gov.hmrc.http.HeaderCarrier
-import uk.gov.hmrc.play.audit.EventKeys._
 import uk.gov.hmrc.play.audit.http.connector.AuditConnector
 import uk.gov.hmrc.play.audit.model.DataEvent
+import uk.gov.hmrc.play.bootstrap.backend.controller.BackendHeaderCarrierProvider
 import uk.gov.hmrc.play.bootstrap.config.{ControllerConfigs, HttpAuditEvent}
-import uk.gov.hmrc.play.bootstrap.filters.{AuditFilter, RequestBodyCaptor, ResponseBodyCaptor}
-import uk.gov.hmrc.play.http.HeaderCarrierConverter
+import uk.gov.hmrc.play.bootstrap.filters.CommonAuditFilter
 
-import scala.concurrent.{ExecutionContext, Future, Promise}
-import scala.util.{Failure, Success, Try}
+import scala.concurrent.ExecutionContext
 
-trait BackendAuditFilter extends AuditFilter {
+trait BackendAuditFilter
+  extends CommonAuditFilter
+     with BackendHeaderCarrierProvider {
 
-  private val logger = Logger(getClass)
+  override protected def filterResponseBody(result: Result, response: ResponseHeader, responseBody: String): String =
+    responseBody
 
-  protected implicit def ec: ExecutionContext
-  def auditConnector: AuditConnector
+  override protected def buildRequestDetails(requestHeader: RequestHeader, request: String): Map[String, String] =
+    Map.empty
 
-  def controllerNeedsAuditing(controllerName: String): Boolean
-
-  def dataEvent(
-    eventType: String,
-    transactionName: String,
-    request: RequestHeader,
-    detail: Map[String, String] = Map()
-  )(implicit
-    hc: HeaderCarrier
-  ): DataEvent
-
-  implicit def mat: Materializer
-
-  val maxBodySize = 32665
-
-  val requestReceived = "RequestReceived"
-
-  def apply(nextFilter: EssentialAction) = new EssentialAction {
-    def apply(requestHeader: RequestHeader) = {
-      val next: Accumulator[ByteString, Result] = nextFilter(requestHeader)
-
-      implicit val hc = HeaderCarrierConverter.fromRequest(requestHeader)
-
-      val loggingContext = s"${requestHeader.method} ${requestHeader.uri}"
-
-      def performAudit(requestBody: String, tryResult: Try[Result])(responseBody: String): Unit = {
-        val detail = tryResult match {
-          case Success(result) =>
-            Map(
-              ResponseMessage -> responseBody,
-              StatusCode -> result.header.status.toString
-            )
-          case Failure(f) =>
-            Map(FailedRequestMessage -> f.getMessage)
-        }
-        auditConnector.sendEvent(
-          dataEvent(
-            eventType       = requestReceived,
-            transactionName = requestHeader.uri,
-            request         = requestHeader,
-            detail          = detail
-          )
-        )
-      }
-
-      if (needsAuditing(requestHeader)) {
-        onCompleteWithInput(loggingContext, next, performAudit)
-      } else next
-    }
-  }
-
-  protected def needsAuditing(request: RequestHeader): Boolean =
-    request.attrs.get(Attrs.HandlerDef).forall { handlerDef =>
-      controllerNeedsAuditing(handlerDef.controller)
-    }
-
-  protected def onCompleteWithInput(
-    loggingContext: String,
-    next: Accumulator[ByteString, Result],
-    handler: (String, Try[Result]) => String => Unit)(
-    implicit ec: ExecutionContext): Accumulator[ByteString, Result] = {
-    val requestBodyPromise = Promise[String]()
-    val requestBodyFuture  = requestBodyPromise.future
-
-    var requestBody: String = ""
-    def callback(body: ByteString): Unit = {
-      requestBody = body.decodeString("UTF-8")
-      requestBodyPromise success requestBody
-    }
-
-    //grabbed from plays csrf filter
-    val wrappedAcc: Accumulator[ByteString, Result] = Accumulator(
-      Flow[ByteString]
-        .via(new RequestBodyCaptor(loggingContext, maxBodySize, callback))
-        .splitWhen(_ => false)
-        .prefixAndTail(0)
-        .map(_._2)
-        .concatSubstreams
-        .toMat(Sink.head[Source[ByteString, _]])(Keep.right)
-    ).mapFuture { bodySource =>
-      next.run(bodySource)
-    }
-
-    wrappedAcc
-      .mapFuture { result =>
-        requestBodyFuture flatMap { res =>
-          {
-            val auditedBody = result.body match {
-              case str: Streamed => {
-                val auditFlow = Flow[ByteString].alsoTo(
-                  new ResponseBodyCaptor(loggingContext, maxBodySize, handler(requestBody, Success(result))))
-                str.copy(data = str.data.via(auditFlow))
-              }
-              case h: HttpEntity => {
-                h.consumeData map { rb =>
-                  val auditString = if (rb.size > maxBodySize) {
-                    logger.warn(
-                      s"txm play auditing: $loggingContext response body ${rb.size} exceeds maxLength $maxBodySize - do you need to be auditing this payload?")
-                    rb.take(maxBodySize).decodeString("UTF-8")
-                  } else {
-                    rb.decodeString("UTF-8")
-                  }
-                  handler(res, Success(result))(auditString)
-                }
-                h
-              }
-            }
-            Future(result.copy(body = auditedBody))
-          }
-        }
-      }
-      .recover[Result] {
-        case ex: Throwable =>
-          handler(requestBody, Failure(ex))("")
-          throw ex
-      }
-  }
+  override protected def buildResponseDetails(response: ResponseHeader): Map[String, String] =
+    Map.empty
 }
 
 class DefaultBackendAuditFilter @Inject()(
@@ -168,16 +47,17 @@ class DefaultBackendAuditFilter @Inject()(
   override val auditConnector: AuditConnector,
   httpAuditEvent: HttpAuditEvent,
   override val mat: Materializer
-)(implicit protected val ec: ExecutionContext) extends BackendAuditFilter {
+)(implicit protected val ec: ExecutionContext
+) extends BackendAuditFilter {
 
   override def controllerNeedsAuditing(controllerName: String): Boolean =
     controllerConfigs.controllerNeedsAuditing(controllerName)
 
   override def dataEvent(
-    eventType: String,
+    eventType      : String,
     transactionName: String,
-    request: RequestHeader,
-    detail: Map[String, String]
+    request        : RequestHeader,
+    detail         : Map[String, String]
   )(implicit hc: HeaderCarrier): DataEvent =
     httpAuditEvent.dataEvent(eventType, transactionName, request, detail)
 }
